+++ sdk/core/System.ClientModel/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Other Changes
 
+- Removed `ReturnWhen` enum in favor of using bool `waitUntilCompleted` parameter in third-party client LRO method signatures.
+
 ## 1.1.0-beta.6 (2024-08-01)
 
 ### Features Added